WIP: Refactoring persistence out of ParamStore
